ミスの修正

+++ src/domain/services/CoincheckRetryTradeService.ts
@@ -29,31 +29,36 @@ export class CoincheckRetryTradeService {
       console.error('オープンオーダーの取得に失敗しました');
       return;
     }
-    openOrders.orders.forEach(async (order) => {
+    for (const order of openOrders.orders) {
       try {
+        // 現在の価格を取得
+        const pairWithoutJpy = order.pair.replace(/_jpy$/i, '');
+        const currentPrice = await this.client.getCurrentPrice(pairWithoutJpy as CoinType);
+        if (order.rate == currentPrice) {
+          console.log(`現在価格と注文価格が一致しているため、再トレードを実行しません: ${order.created_at}　現在価格: ${currentPrice}　注文価格: ${order.rate}`);
+          continue;
+        }
         // まずは注文をキャンセル
         const deleteResult = await this.client.deleteOpenOrder(order.id);
         if (deleteResult.success === false) {
           console.error('オープンオーダーのキャンセルに失敗しました', deleteResult);
-          return;
+          continue;
         }
-        // 現在の価格を取得
-        const pairWithoutJpy = order.pair.replace(/_jpy$/i, '');
-        const currentPrice = await this.client.getCurrentPrice(pairWithoutJpy as CoinType);
+        // 注文キャンセル後、1秒待機
+        await new Promise(resolve => setTimeout(resolve, 1000));
         // 現在の所持コイン量を取得
         const amount = await this.client.getCoinBalance(pairWithoutJpy as CoinType);
         // 再トレードを実行
         await this.client.createOrder({
           rate: currentPrice,
           amount: order.order_type === "sell" ? amount : Number(order.pending_amount),
-          
           order_type: order.order_type,
           pair: order.pair
         });
-        console.log(`再トレード実行: ${order.created_at}`);
+        console.log(`再トレード実行: ${order.pair}${order.created_at}`);
       } catch (error) {
         console.error('再トレードの実行に失敗しました', error);
       }
-    });
+    };
   }
 }
